Reference

ParticleField

FLOWVPM.ParticleField — Type

`ParticleField(maxparticles::Int, R=FLOAT_TYPE; <keyword arguments>)`

Create a new particle field with maxparticles particles. The particle field is created with the default values for the other parameters.

Arguments

maxparticles::Int : Maximum number of particles in the field.

R=FLOAT_TYPE : Type of the particle field. Default is FLOAT_TYPE.

formulation : VPM formulation. Default is rVPM.

viscous::ViscousScheme : Viscous scheme. Default is Inviscid(). With rVPM formulation, a viscous scheme is not required for numerical stability.

np::Int : Number of particles currently in the field. Default is 0. (user should not modify)

nt::Int : Current time step number. Default is 0. (user should not modify)

t::Real : Current time. Default is 0.

transposed::Bool : If true, the transposed scheme is recommended for stability. Default is true. (user should not modify)

fmm::FMM{TEPS} : Fast-multipole settings. Default is FMM().

M::Array{R, 1} : Memory for computations. Default is zeros(R, 4). (user should not modify)

toggle_rbf::Bool : If true, the FMM computes the vorticity field rather than velocity field. This is used as an internal switch for the FMM. Default is false. (user should not modify)

toggle_sfs::Bool : If true, the FMM computes the stretching term for the SFS model. This is used as an internal switch for the FMM. Default is false. (user should not modify)

SFS::S : Subfilter-scale contributions scheme. Default is noSFS.

kernel::Tkernel : Vortex particle kernel. Default is gaussianerf.

UJ::TUJ : Particle-to-particle calculation. Default is UJ_fmm.

Uinf::TUinf : Uniform freestream function Uinf(t). Default is no freestream.

relaxation::TR : Relaxation scheme. Default is pedrizzetti.

integration::Tintegration : Time integration scheme. Default is rungekutta3. The only other option is euler.

useGPU::Int : Run on GPU if >0, CPU if 0. Default is 0. (Experimental and does not accelerate SFS calculations)

Formulations

FLOWVPM.rVPM — Constant

rVPM

Alias for the reformulated VPM formulation. Enforces conservation of mass and momentum for a spherical fluid element and is the default formulation (f=0, g=1/5)

FLOWVPM.cVPM — Constant

cVPM

Alias for the classic VPM formulation.

FLOWVPM.formulation_tube_continuity — Constant

formulation_tube_continuity

Alias for mass conserving tube formulation. Enforces conservation of mass for a vortex tube (f=1/2, g=0)

FLOWVPM.formulation_tube_momentum — Constant

formulation_tube_momentum

Alias for momentum conserving tube formulation. Enforces conservation of momentum for a vortex tube (f=1/4, g=1/4)

Viscous Models

FLOWVPM.Inviscid — Type

Inviscid()

Creates an inviscid scheme with zero kinematic viscosity.

FLOWVPM.CoreSpreading — Type

`CoreSpreading(nu, sgm0, zeta::Tzeta=zeta_fmm; <keyword arguments>)`

Creates a core spreading viscous scheme with the given parameters.

Arguments

nu::Real Kinematic viscosity.

sgm0::Real Core size after reset.

zeta::Function = zeta_fmm Basis function evaluation method.

beta::Real = 1.5 Maximum core size growth σ/σ_0.

itmax::Int = 15 Maximum number of RBF iterations.

tol::Real = 1e-3 RBF interpolation tolerance.

iterror::Bool = true Throw error if RBF didn't converge.

verbose::Bool = false Verbose on RBF interpolation.

v_lvl::Int = 2 Verbose printing tab level.

debug::Bool = false Print verbose for debugging.

rbf::Function = rbf_conjugategradient RBF function.

rr0s::Array{R, 1} = zeros(R, 3) Initial field residuals.

rrs::Array{R, 1} = zeros(R, 3) Current field residuals.

prev_rrs::Array{R, 1} = zeros(R, 3) Previous field residuals.

pAps::Array{R, 1} = zeros(R, 3) pAp product.

alphas::Array{R, 1} = zeros(R, 3) Alpha coefficients.

betas::Array{R, 1} = zeros(R, 3)

flags::Array{Bool, 1} = zeros(Bool, 3)

FLOWVPM.ParticleStrengthExchange — Type

`ParticleStrengthExchange(nu; <keyword arguments>)`

Creates a particle strength exchange viscous scheme with the given parameters.

Arguments

nu::Real Kinematic viscosity.

recalculate_vols::Bool = true Whether to recalculate particle volumes.

FMM

FLOWVPM.FMM — Type

`FMM(; p::Int=4, ncrit::Int=50, theta::Real=0.4, phi::Real=0.3)`

Parameters for FMM solver.

Arguments

p : Order of multipole expansion (number of terms).

ncrit : Maximum number of particles per leaf.

theta : Neighborhood criterion. This criterion defines the distance where the far field starts. The criterion is that if θ*r < R1+R2 the interaction between two cells is resolved through P2P, where r is the distance between cell centers, and R1 and R2 are each cell radius. This means that at θ=1, P2P is done only on cells that have overlap; at θ=0.5, P2P is done on cells that their distance is less than double R1+R2; at θ=0.25, P2P is done on cells that their distance is less than four times R1+R2; at θ=0, P2P is done on cells all cells.

phi : Regularizing neighborhood criterion. This criterion avoid approximating interactions with the singular-FMM between regularized particles that are sufficiently close to each other across cell boundaries. Used together with the θ-criterion, P2P is performed between two cells if φ < σ/dx, where σ is the average smoothing radius in between all particles in both cells, and dx is the distance between cell boundaries ( dx = r-(R1+R2) ). This means that at φ = 1, P2P is done on cells with boundaries closer than the average smoothing radius; at φ = 0.5, P2P is done on cells closer than two times the smoothing radius; at φ = 0.25, P2P is done on cells closer than four times the smoothing radius.

Subfilter-Scale Models

FLOWVPM.noSFS — Constant

noSFS

Alias for the no subfilter-scale model.

FLOWVPM.SFS_Cs_nobackscatter — Constant

SFS_Cs_nobackscatter

Alias for the Constant SFS model with no backscatter.

FLOWVPM.SFS_Cd_twolevel_nobackscatter — Constant

SFS_Cd_twolevel_nobackscatter

Alias for the Dynamic SFS model with two levels and no backscatter. This is the recommended SFS model for high fidelity modeling.

FLOWVPM.SFS_Cd_threelevel_nobackscatter — Constant

SFS_Cd_threelevel_nobackscatter

Alias for the Dynamic SFS model with three levels and no backscatter. This is similar to the two level version but uses a lower value of alpha (0.667).

Particle Kernels

FLOWVPM.singular — Constant

singular

Alias for the singular kernel.

FLOWVPM.gaussian — Constant

gaussian

Alias for the Gaussian kernel.

FLOWVPM.gaussianerf — Constant

gaussianerf

Alias for the Gaussian error function kernel.

FLOWVPM.winckelmans — Constant

winckelmans

Alias for the Winckelmans kernel.

Relaxation Schemes

FLOWVPM.norelaxation — Constant

norelaxation

Alias for the no relaxation scheme.

FLOWVPM.pedrizzetti — Constant

pedrizzetti

Alias for the Pedrizzetti relaxation scheme.

FLOWVPM.correctedpedrizzetti — Constant

correctedpedrizzetti

Alias for the corrected Pedrizzetti relaxation scheme. Is a modification to the pedrizzetti relaxation that preserves the vortex strength magnitude.

Time Integration

FLOWVPM.euler — Function

euler(pfield, dt; relax=false)

Convects the ::ParticleField by timestep dt using a forward Euler step.

FLOWVPM.rungekutta3 — Function

Steps the field forward in time by dt in a third-order low-storage Runge-Kutta integration scheme. See Notebook entry 20180105.

rungekutta3(pfield::ParticleField, dt::R3; relax::Bool=false, custom_UJ=nothing)

Steps the field forward in time by dt in a third-order low-storage Runge-Kutta integration scheme using the VPM reformulation. See Notebook entry 20180105 (RK integration) and notebook 20210104 (reformulation).